fix(build): handle test functions better
Currently, I just build the test functions I want to run and comment out
the rest. This is not a good way to do it. Improve on it by moving the
test functions to a separate file and pass the test functions as an
argument to make instead. Note, these are just "scratch" functions I use
to test isolated parts of the code during development.
