+++ .github/workflows/ci.yml
@@ -10,6 +10,6 @@ jobs:
             - run: make format && git diff --quiet
             - run: TOOLS_PATH=/home/ubuntu/dev/tools make HW=NSUMO
             - run: TOOLS_PATH=/home/ubuntu/dev/tools make HW=LAUNCHPAD
+            - run: TOOLS_PATH=/home/ubuntu/dev/tools tools/build_tests.sh
             - run: make cppcheck
-
