+++ Makefile
@@ -1,8 +1,8 @@
 # Check arguments
 ifeq ($(HW),LAUNCHPAD) #HW argument
-TARGET_NAME = launchpad
+TARGET_HW = launchpad
 else ifeq ($(HW),NSUMO)
-TARGET_NAME = nsumo
+TARGET_HW = nsumo
 else ifeq ($(MAKECMDGOALS),clean)
 else ifeq ($(MAKECMDGOALS),cppcheck)
 else ifeq ($(MAKECMDGOALS),format)
@@ -10,14 +10,22 @@ else ifeq ($(MAKECMDGOALS),format)
 else
 $(error "Must pass HW=LAUNCHPAD or HW=NSUMO")
 endif
+TARGET_NAME = $(TARGET_HW)
 
+ifneq ($(TEST),) # TEST argument
+ifeq ($(findstring test_,$(TEST)),)
+$(error "TEST=$(TEST) is invalid (test functions must start with test_)")
+else
+TARGET_NAME = $(TEST)
+endif
+endif
 # Directories
 TOOLS_DIR = ${TOOLS_PATH}
 MSPGCC_ROOT_DIR =$(TOOLS_DIR)/msp430-gcc
 MSPGCC_BIN_DIR = $(MSPGCC_ROOT_DIR)/bin
 MSPGCC_INCLUDE_DIR = $(MSPGCC_ROOT_DIR)/include
-BUILD_DIR = build/$(TARGET_NAME)
-OBJ_DIR = $(BUILD_DIR)/obj
+BUILD_DIR = build
+OBJ_DIR = $(BUILD_DIR)/obj/$(TARGET_HW)
 TI_CCS_DIR = $(TOOLS_DIR)/ccs1250/ccs
 DEBUG_BIN_DIR = $(TI_CCS_DIR)/ccs_base/DebugServer/bin
 DEBUG_DRIVERS_DIR = $(TI_CCS_DIR)/ccs_base/DebugServer/drivers
@@ -36,7 +44,7 @@ CPPCHECK = cppcheck
 FORMAT = clang-format-12
 
 # Files
-TARGET = $(BUILD_DIR)/$(TARGET_NAME)
+TARGET = $(BUILD_DIR)/bin/$(TARGET_HW)/$(TARGET_NAME)
 
 SOURCES_WITH_HEADERS = \
 			 src/common/assert_handler.c \
@@ -46,14 +54,17 @@ SOURCES_WITH_HEADERS = \
 			 src/app/drive.c  \
 			 src/app/enemy.c \
 
-TEST_SRC = $(addprefix src/test/, \
-						test.c)	
-
+ifdef TEST
+SOURCES = 	\
+			src/test/test.c \
+			$(SOURCES_WITH_HEADERS)
+# Delete object file to force rebuild when changing test (there is probably a better way ...)
+$(shell rm -f $(BUILD_DIR)/obj/src/test/test.o)
+else
 SOURCES = \
 		src/main.c \
-		$(SOURCES_WITH_HEADERS) \
-		$(TEST_SRC) \
-
+		$(SOURCES_WITH_HEADERS)
+endif
 HEADERS = \
 		$(SOURCES_WITH_HEADERS:.c=.h) \
 		src/common/defines.h \
@@ -63,7 +74,9 @@ OBJECTS = $(patsubst %,$(OBJ_DIR)/%,$(OBJECT_NAMES))
 
 # Defines
 HW_DEFINE = $(addprefix -D,$(HW))
-DEFINES = $(HW_DEFINE)
+TEST_DEFINES = $(addprefix -DTEST=,$(TEST))
+DEFINES = $(HW_DEFINE) \
+		  $(TEST_DEFINES)
 
 # Static Analysis
 ## Don't check the msp430 helper headers (they have a lot of ifdefs)
